requirements.txt
# Python Dependencies for YojnaSathi Backend
# Install these using: pip install -r requirements.txt

# FastAPI - Modern web framework for building APIs
fastapi>=0.104.1
starlette>=0.27.0

# Uvicorn - ASGI server to run FastAPI application
uvicorn[standard]>=0.24.0

# Pydantic - Data validation using Python type annotations
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Email validation
email-validator>=2.0.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
passlib>=1.7.4
cryptography>=41.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Utilities
python-multipart>=0.0.6
typing-extensions>=4.8.0
